simplify retry logic
