--- app/lib/retry_exception.rb
@@ -1,2 +1,0 @@
-# Exception for mysql retries
-class RetryException < RuntimeError; end
